hi655x-pmic: Fixup issue with un-acked interrupts
While trying to get the powerkey to funciton, I found
when pressing the key, I would get infinitely repeating
interrupts.

After digging around a bit, it seems we didn't set the
ack_base value for the regmap irqchip logic, so nothing
was acking the interrupt.

This patch adds the ack_base, which seems to make things
work.

Signed-off-by: John Stultz <john.stultz@linaro.org>