HACK: adv7511: Enable HPD interrupts to avoid monitor detection issues
still not entirely clear about it, but it looks like we are just lucky in
the 4.4 kernels to not access the STATUS register at the wrong time.
I printed the STATUS register at several places in the driver, and it
reflected the wrong HPD status at times in the 4.4 kernel too.

One more difference in the 4.4 driver is that hotplug detect interrupts
are enabled, which is missing in the upstream driver. For some reason,
adding that to the upstream driver seems to get the STATUS register behave
correct every time we try to read it. HPD interrupt changes shared below.

I'll spend some time trying to root cause this behaviour. If I can't
figure it out, I'll just post a patch to enable HPD interrupts.

Please let me know how this works on Hikey boards.

Thanks,
Archit

Signed-off-by: John Stultz <john.stultz@linaro.org>