arm64: dma-mapping: Use swiotlb ops for non-device operator
Fix cache interface for default NULL device

It is commit
commit 1dccb598df549d892b6450c261da54cdd7af44b4
Author: Arnd Bergmann <arnd@arndb.de>
Date:   Mon Nov 16 17:25:48 2015 +0100

    arm64: simplify dma_get_ops

This commit will let all dma-ops have a device.

In ion module:
ion_sync_for_device
                dma_sync_sg_for_device(NULL, buffer->sg_table->sgl,
                               buffer->sg_table->nents, DMA_BIDIRECTIONAL);

The first parameter is Null. The ion code wants to use this interface to
do cache maintaince. But with the change of arm64 DMA-API change,here is
a dummy operator.

Signed-off-by: Chen Feng <puck.chen@hisilicon.com>