clk: hisilicon: Hi3660: correct clock id
In the old code only allocates one clock id for total 4 clocks, as
result when we refer wrongly to DDR clock when execute CPU frequency
change. This patch is to assign dedicated clock id for every clock, so
finally we can use clock node + clock id in DT to bind CPU clock
correctly.

Change-Id: I99ad504c269a35ff154fddddd1e743c25f41b555
Signed-off-by: Leo Yan <leo.yan@linaro.org>