arm64: dts: hi3660: Lower sustained-power to 4000
I ran a 100% duty-cycle rt-app workload on 4 big cores for 10 minutes.
This leads the platform to heat to temperatures above 65°C, so IPA is
active during the whole duration of the workload. As such, we can
monitor how well IPA regulates the temperature - it should stay as
close to 75°C (control temperature) as possible. The absolute difference
between the measured temperature and the control temperature is thus an
error value that can be measured and compared between runs.

These are the results of 10 minutes of workload per sustained-power
value:

power   error mean      error median    error min   error max

4500    3684.315836     3636.0          49          11104
4250    3295.916754     3022.0          49          9572
4000    3186.842631     3022.0          49          11513
3750    3512.759481     3840.0          49          9982

While the temperature evolution curve doesn't seem to change
drastically, lowing the sustained-power value to 4000 does slightly
reduce the temperature oscillations, leading to a slightly better
temperature stability.

The average error does increase if the sustained-power value is too
low however, as IPA will limit the platform too much and the temperature
will take more time to ramp up to the control temperature.

Change-Id: I416e2a99c398ae2547d73c7831e4b71d3b5a86ce
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>