arm64: dts: hisi: Update pcie node
PCIe node was copied from Linux mainline. However, this is wrong. Driver
used in this kernel (v4.9) is different from linux mainline. Acutually
PCIe driver in v4.9 is more like the one in v4.4 kernel. So, PCIe node
should correspond to that.

As a correction, in this commit:
 - Update the value of compatible from "hisilicon,kirin960-pcie" to
   "hisilicon,kirin-pcie".
 - Update the property name from "reset-gpios" to "reset-gpio".

Without this patch, PCIe dts cannot find the matching driver.

Cc: Guodong Xu <guodong.xu@linaro.org>
Signed-off-by: Loic Poulain <loic.poulain@linaro.org>
Signed-off-by: Xiaowei Song <songxiaowei@hisilicon.com>