arm64: dts: hi3660: Update sched-energy values
This is the result of a combined effort with Leo to try and improve
the energy model of the HiKey960. Each OPP has been benchmarked with
dhry2.1 while the board was cooled with a heatsink and a fan to
prevent any thermal mitigation. The DDR frequency was also locked at
400MHz. The results show that the A53 capacity values can be slightly
increased.

Idle costs are trickier to analyse on the HiKey960, and Leo rightfully
pointed out that they should not be greater than active costs - this
is actually the case with the current values, as the A73 cluster idle
costs are greater than the lowest cluster active costs.

As such, each idle value is defined as:
min(measured_value, min(active_costs))

Change-Id: Ibae97aeb9beee60c29ebfc7a0573490e06a4ec09
Co-Authored-By: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>
(cherry picked from commit b867b9875035fa040f89405f4d57a429fd0bd0d5)
Signed-off-by: Valentin Schneider <valentin.schneider@arm.com>