arm64: dts: hi3660: use the simple power model for the GPU
The Mali driver has an improved power model for Mali-G71 based on
performance counters.  Reading performance counters requires the GPU to
be powered on continuously.  Currently, the driver does not support
disabling the performance counters when the GPU goes idle.  This
interferes with the code that powers off the GPU, which keeps firing a
timer every 400us waiting for the GPU to power off, which never happens
(see bug https://bugs.96boards.org/show_bug.cgi?id=652).  This timer
keeps waking the CPUs unnecessarily, wasting power.

The fix for the driver is likely to be pretty complex.  We may get a
fix in the future, via a new version of the driver.  In the meantime, we
need a solution that allows us the thermal core to throttle the GPU via
devfreq and removes the spurious wakeups.  This patch chooses the simple
power model, which is based on utilization and is less accurate but
allows us to keep DEVFREQ_THERMAL=y.

This fixes https://bugs.96boards.org/show_bug.cgi?id=652

Change-Id: Iee14c77fdcd3ab6eaed9e8f955dc8d585bc20897
Signed-off-by: Javi Merino <javi.merino@arm.com>