cpufreq: cpufreq-dt: fix NETLINK wake up during suspend
during suspend, non-boot cpu will be pluged off, and when
the last cpu of big cluster be pluged off, the related cpu
cooling device will be unregistered, it will trigger an
uevent, this will cause early wakeup if there are not enough
time for the uevent to be handled before suspend.

This patch fixes the issue by not unregistering cpu cooling
device during suspend, this may not be a perfect solution.

Signed-off-by: Kevin Wangtao <kevin.wangtao@hisilicon.com>